fix(billing): change EmitirDteRequest cantidad to BigDecimal to prevent Jackson parse error when receiving decimals from POS

+++ backend/billing-service/src/main/java/com/poscl/billing/api/dto/EmitirDteRequest.java
@@ -103,7 +103,7 @@ public class EmitirDteRequest {
 
         @NotNull(message = "La cantidad es requerida")
         @Positive(message = "La cantidad debe ser positiva")
-        private Integer cantidad;
+        private BigDecimal cantidad;
 
         private BigDecimal montoTotal;
